backend/requirements.txt
openai
requests
google-api-python-client
SpeechRecognition
